refactor: export useful checkpoint types and class (#33)
This updates the index file for checkpoint to ensure it exports useful
types and classes, so it is easily accessible to potential users.

It also cleans up the sx-api server code to use only the checkpoint
entrypoint and not reach down into non exported subfolders.
